chore: simplified UI folder structure PSD-3 (#69)
