+++ client/.gitignore
@@ -42,4 +42,6 @@ custom-elements.json
 # User files that may contain sensitive tokens etc.
 .npmrc
 
-bootstrapDone
+# Gradle tasks
+.bootstrapDone
+.tests.executed
