+++ client/Dockerfile
@@ -5,8 +5,6 @@ SHELL ["/bin/bash", "-c"]
 WORKDIR /app
 
 COPY package.json /app/package.json
-COPY .build /app/.build
-COPY web/package.json /app/web/package.json
 
 COPY . /app
