+++ client/public/initSSO.js
