+++ client/settings.gradle.kts
@@ -1,3 +1,1 @@
 rootProject.name = "client"
-
-include("web")
