+++ client/src/main/index.ts
