+++ client/src/routes/not-found/not-found.styles.ts
