+++ client/src/routes/not-found/not-found.template.ts
