--- client/web/.gitignore
@@ -1,43 +1,0 @@
-# Client specific
-
-# Logs
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Output
-dist/
-
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov
-
-# Coverage directory used by tools like istanbul
-coverage
-
-# Coverage files
-coverage.lcov
-
-# nyc test coverage
-.nyc_output
-
-# Optional npm cache directory
-.npm
-
-# Optional eslint cache
-.eslintcache
-
-# Optional REPL history
-.node_repl_history
-
-# lerna backups
-package.json.lerna_backup
-
-# custom elements metadata
-custom-elements.json
-
-# TypeScript incremental compilation cache
-*.tsbuildinfo
-.rollup.cache
-
-# User files that may contain sensitive tokens etc.
-.npmrc
