--- client/web/package.json
@@ -1,96 +1,0 @@
-{
-  "name": "{{pkgName}}",
-  "description": "",
-  "version": "{{applicationVersionWeb}}",
-  "private": true,
-  "license": "UNLICENSED",
-  "config": {
-    "API_HOST": "{{apiHost}}",
-    "PORT": 6060,
-    "ENABLE_SSO": {{enableSSO}}
-  },
-  "genx": {
-    "app": {
-      "rootElement": "{{rootElement}}"
-    },
-    "federation": {
-      "remotes": [
-        "@genesislcap/foundation-zero"
-      ]
-    }
-  },
-  "scripts": {
-    "baseline": "npm run clean && npm run bootstrap",
-    "bootstrap": "npm install --no-fund --no-audit",
-    "bootstrap:ci": "npm ci --no-fund --no-audit",
-    "build": "genx build -e ENABLE_SSO",
-    "build:stats": "genx analyze",
-    "clean": "genx clean dist node_modules",
-    "dev": "genx dev -e API_HOST,ENABLE_SSO",
-    "dev:docker": "npm run dev -- --host 0.0.0.0",
-    "dev:intellij": "genx dev -e ENABLE_SSO",
-    "dev:no-open": "npm run dev -- --no-open",
-    "dev:https": "npm run dev -- --https",
-    "dev:webpack": "npm run dev -- -b webpack",
-    "dsconfig": "dsconfig --path src/styles/design-tokens.json",
-    "git:setup": "cd ../.. && npx --yes husky install",
-    "lint": "genx lint --profile",
-    "lint:fix": "genx lint --fix",
-    "lint:eslint": "genx lint -l eslint --profile",
-    "lint:stylelint": "genx lint -l stylelint",
-    "serve": "genx serve",
-    "test": "genx test",
-    "test:coverage": "genx test --coverage",
-    "test:unit:watch": "genx test --watch"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "*.{ts,js}": [
-      "eslint"
-    ],
-    "*.styles.ts": [
-      "stylelint"
-    ]
-  },
-  "eslintConfig": {
-    "extends": "@genesislcap/eslint-config"
-  },
-  "stylelint": {
-    "extends": "@genesislcap/stylelint-config"
-  },
-  "prettier": "@genesislcap/prettier-config",
-  "devDependencies": {
-    "@genesislcap/design-system-configurator": "14.67.4",
-    "@genesislcap/foundation-testing": "14.67.4",
-    "@genesislcap/genx": "14.67.4",
-    "esm": "^3.2.25",
-    "husky": "^7.0.4",
-    "lint-prepush": "^2.2.1",
-    "lint-staged": "^12.4.1"
-  },
-  "dependencies": {
-    "@genesislcap/foundation-comms": "14.67.4",
-    "@genesislcap/foundation-entity-management": "14.67.4",
-    "@genesislcap/foundation-forms": "14.67.4",
-    "@genesislcap/foundation-header": "14.67.4",
-    "@genesislcap/foundation-inbox": "14.67.4",
-    "@genesislcap/foundation-layout": "14.67.4",
-    "@genesislcap/foundation-login": "14.67.4",
-    "@genesislcap/foundation-ui": "14.67.4",
-    "@genesislcap/foundation-utils": "14.67.4",
-    "@genesislcap/foundation-zero": "14.67.4",
-    "@genesislcap/foundation-zero-grid-pro": "14.67.4",
-    "@genesislcap/g2plot-chart": "14.67.4",
-    "@microsoft/fast-components": "^2.21.6",
-    "@microsoft/fast-element": "^1.7.0",
-    "@microsoft/fast-foundation": "^2.33.4",
-    "@microsoft/fast-router": "^0.4.2",
-    "@microsoft/fast-web-utilities": "^5.1.0",
-    "rxjs": "^7.5.4",
-    "tslib": "^2.3.1"
-  }
-}
